+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Change Log
 25. Removed view:animated event (1.8.4) use 'show' instead;
 26. Added dnd support to views (drag, drop, sortable);
 27. Added selectable item support to views (selectable);
+28. Refined template caching (local-dom, remote and in-memory);
 
 working on ...